requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-restful
flask_security
email_validator
bcrypt
numpy
matplotlib

